tests/requirements.txt
coverage
django-app-helper
djangocms-text-ckeditor
tox
